fix(01-09): test_plugin AbiError.code uses AbiErrorCode directly
- Remove u32 casts from AbiErrorCode::Generic in polyplug_init
- AbiError.code field is AbiErrorCode enum type, not u32

+++ tests/fixtures/test_plugin/src/lib.rs
@@ -108,13 +108,13 @@ pub unsafe extern "C" fn polyplug_init(
 ) -> AbiError {
     if host_vtable.is_null() {
         return AbiError {
-            code: AbiErrorCode::Generic as u32,
+            code: AbiErrorCode::Generic,
             message: string_view_null(),
         };
     }
     if ctx.is_null() {
         return AbiError {
-            code: AbiErrorCode::Generic as u32,
+            code: AbiErrorCode::Generic,
             message: string_view_null(),
         };
     }
